reorder configs
