+++ src/main/java/de/hysky/skyblocker/config/SkyblockerConfig.java
@@ -4,42 +4,42 @@ import de.hysky.skyblocker.config.configs.*;
 import dev.isxander.yacl3.config.v2.api.SerialEntry;
 
 public class SkyblockerConfig {
-	@SerialEntry
-	public int version = 2;
+    @SerialEntry
+    public int version = 2;
 
-	@SerialEntry
-	public GeneralConfig general = new GeneralConfig();
+    @SerialEntry
+    public GeneralConfig general = new GeneralConfig();
 
-	@SerialEntry
-	public UiAndVisualsConfig uiAndVisuals = new UiAndVisualsConfig();
+    @SerialEntry
+    public UIAndVisualsConfig uiAndVisuals = new UIAndVisualsConfig();
 
-	@SerialEntry
-	public HelperConfig helper = new HelperConfig();
+    @SerialEntry
+    public HelperConfig helper = new HelperConfig();
 
-	@SerialEntry
-	public DungeonsConfig dungeons = new DungeonsConfig();
+    @SerialEntry
+    public DungeonsConfig dungeons = new DungeonsConfig();
 
-	@SerialEntry
-	public CrimsonIsleConfig crimsonIsle = new CrimsonIsleConfig();
+    @SerialEntry
+    public CrimsonIsleConfig crimsonIsle = new CrimsonIsleConfig();
 
-	@SerialEntry
-	public MiningConfig mining = new MiningConfig();
+    @SerialEntry
+    public MiningConfig mining = new MiningConfig();
 
-	@SerialEntry
-	public FarmingConfig farming = new FarmingConfig();
+    @SerialEntry
+    public FarmingConfig farming = new FarmingConfig();
 
-	@SerialEntry
-	public OtherLocationsConfig otherLocations = new OtherLocationsConfig();
+    @SerialEntry
+    public OtherLocationsConfig otherLocations = new OtherLocationsConfig();
 
-	@SerialEntry
-	public SlayerConfig slayer = new SlayerConfig();
+    @SerialEntry
+    public SlayersConfig slayer = new SlayersConfig();
 
-	@SerialEntry
-	public MessagesConfig messages = new MessagesConfig();
+    @SerialEntry
+    public MessagesFilterConfig messages = new MessagesFilterConfig();
 
-	@SerialEntry
-	public QuickNavConfig quickNav = new QuickNavConfig();
+    @SerialEntry
+    public QuickNavigationConfig quickNav = new QuickNavigationConfig();
 
-	@SerialEntry
-	public MiscConfig misc = new MiscConfig();
+    @SerialEntry
+    public MiscConfig misc = new MiscConfig();
 }
