+++ src/main/java/de/hysky/skyblocker/config/configs/QuickNavigationConfig.java
@@ -2,7 +2,7 @@ package de.hysky.skyblocker.config.configs;
 
 import dev.isxander.yacl3.config.v2.api.SerialEntry;
 
-public class QuickNavConfig {
+public class QuickNavigationConfig {
     @SerialEntry
     public boolean enableQuickNav = true;
