+++ src/main/java/de/hysky/skyblocker/skyblock/fancybars/FancyStatusBars.java
@@ -2,9 +2,8 @@ package de.hysky.skyblocker.skyblock.fancybars;
 
 import com.google.gson.JsonObject;
 import de.hysky.skyblocker.SkyblockerMod;
-import de.hysky.skyblocker.config.SkyblockerConfig;
 import de.hysky.skyblocker.config.SkyblockerConfigManager;
-import de.hysky.skyblocker.config.configs.UiAndVisualsConfig;
+import de.hysky.skyblocker.config.configs.UIAndVisualsConfig;
 import de.hysky.skyblocker.skyblock.StatusBarTracker;
 import de.hysky.skyblocker.utils.Utils;
 import de.hysky.skyblocker.utils.scheduler.Scheduler;
@@ -59,7 +58,7 @@ public class FancyStatusBars {
         // Fetch from old status bar config
         int[] counts = new int[3]; // counts for RIGHT, LAYER1, LAYER2
         StatusBar health = statusBars.get("health");
-        UiAndVisualsConfig.OldBarPositions barPositions = SkyblockerConfigManager.get().uiAndVisuals.bars.barPositions;
+        UIAndVisualsConfig.OldBarPositions barPositions = SkyblockerConfigManager.get().uiAndVisuals.bars.barPositions;
         loadOldBarPosition(health, counts, barPositions.healthBarPosition);
         StatusBar intelligence = statusBars.get("intelligence");
         loadOldBarPosition(intelligence, counts, barPositions.manaBarPosition);
@@ -104,7 +103,7 @@ public class FancyStatusBars {
      * @param counts the counts for each bar position (LAYER1, LAYER2, RIGHT)
      * @param position the position to load
      */
-    private static void loadOldBarPosition(StatusBar bar, int[] counts, UiAndVisualsConfig.OldBarPosition position) {
+    private static void loadOldBarPosition(StatusBar bar, int[] counts, UIAndVisualsConfig.OldBarPosition position) {
         switch (position) {
             case RIGHT:
                 bar.anchor = BarPositioner.BarAnchor.HOTBAR_RIGHT;
