+++ src/main/java/de/hysky/skyblocker/utils/render/title/TitleContainer.java
@@ -1,8 +1,7 @@
 package de.hysky.skyblocker.utils.render.title;
 
-import de.hysky.skyblocker.config.SkyblockerConfig;
 import de.hysky.skyblocker.config.SkyblockerConfigManager;
-import de.hysky.skyblocker.config.configs.UiAndVisualsConfig;
+import de.hysky.skyblocker.config.configs.UIAndVisualsConfig;
 import de.hysky.skyblocker.events.HudRenderEvents;
 import de.hysky.skyblocker.utils.scheduler.Scheduler;
 import net.fabricmc.fabric.api.client.command.v2.ClientCommandManager;
@@ -94,8 +93,8 @@ public class TitleContainer {
         float scale = 3F * (SkyblockerConfigManager.get().uiAndVisuals.titleContainer.titleContainerScale / 100F);
 
         // Grab direction and alignment values
-        UiAndVisualsConfig.Direction direction = SkyblockerConfigManager.get().uiAndVisuals.titleContainer.direction;
-        UiAndVisualsConfig.Alignment alignment = SkyblockerConfigManager.get().uiAndVisuals.titleContainer.alignment;
+        UIAndVisualsConfig.Direction direction = SkyblockerConfigManager.get().uiAndVisuals.titleContainer.direction;
+        UIAndVisualsConfig.Alignment alignment = SkyblockerConfigManager.get().uiAndVisuals.titleContainer.alignment;
         // x/y refer to the starting position for the text
         // y always starts at yPos
         float x = 0;
@@ -107,8 +106,8 @@ public class TitleContainer {
             width += textRenderer.getWidth(title.getText()) * scale + 10;
         }
 
-        if (alignment == UiAndVisualsConfig.Alignment.MIDDLE) {
-            if (direction == UiAndVisualsConfig.Direction.HORIZONTAL) {
+        if (alignment == UIAndVisualsConfig.Alignment.MIDDLE) {
+            if (direction == UIAndVisualsConfig.Direction.HORIZONTAL) {
                 //If middle aligned horizontally, start the xPosition at half of the width to the left.
                 x = xPos - (width / 2);
             } else {
@@ -116,7 +115,7 @@ public class TitleContainer {
                 x = xPos;
             }
         }
-        if (alignment == UiAndVisualsConfig.Alignment.LEFT || alignment == UiAndVisualsConfig.Alignment.RIGHT) {
+        if (alignment == UIAndVisualsConfig.Alignment.LEFT || alignment == UIAndVisualsConfig.Alignment.RIGHT) {
             //If left or right aligned, start at xPos, we will shift each text later
             x = xPos;
         }
@@ -125,14 +124,14 @@ public class TitleContainer {
 
             //Calculate which x the text should use
             float xToUse;
-            if (direction == UiAndVisualsConfig.Direction.HORIZONTAL) {
-                xToUse = alignment == UiAndVisualsConfig.Alignment.RIGHT ?
+            if (direction == UIAndVisualsConfig.Direction.HORIZONTAL) {
+                xToUse = alignment == UIAndVisualsConfig.Alignment.RIGHT ?
                         x - (textRenderer.getWidth(title.getText()) * scale) : //if right aligned we need the text position to be aligned on the right side.
                         x;
             } else {
-                xToUse = alignment == UiAndVisualsConfig.Alignment.MIDDLE ?
+                xToUse = alignment == UIAndVisualsConfig.Alignment.MIDDLE ?
                         x - (textRenderer.getWidth(title.getText()) * scale) / 2 : //if middle aligned we need the text position to be aligned in the middle.
-                        alignment == UiAndVisualsConfig.Alignment.RIGHT ?
+                        alignment == UIAndVisualsConfig.Alignment.RIGHT ?
                                 x - (textRenderer.getWidth(title.getText()) * scale) : //if right aligned we need the text position to be aligned on the right side.
                                 x;
             }
@@ -157,13 +156,13 @@ public class TitleContainer {
             context.getMatrices().pop();
 
             //Calculate the x and y positions for the next title
-            if (direction == UiAndVisualsConfig.Direction.HORIZONTAL) {
-                if (alignment == UiAndVisualsConfig.Alignment.MIDDLE || alignment == UiAndVisualsConfig.Alignment.LEFT) {
+            if (direction == UIAndVisualsConfig.Direction.HORIZONTAL) {
+                if (alignment == UIAndVisualsConfig.Alignment.MIDDLE || alignment == UIAndVisualsConfig.Alignment.LEFT) {
                     //Move to the right if middle or left aligned
                     x += textRenderer.getWidth(title.getText()) * scale + 10;
                 }
 
-                if (alignment == UiAndVisualsConfig.Alignment.RIGHT) {
+                if (alignment == UIAndVisualsConfig.Alignment.RIGHT) {
                     //Move to the left if right aligned
                     x -= textRenderer.getWidth(title.getText()) * scale + 10;
                 }
