+++ src/main/resources/assets/skyblocker/lang/en_us.json
@@ -21,6 +21,9 @@
 
   "skyblocker.title": "Skyblocker Settings",
 
+  "skyblocker.category.useinterfaceandvisuals": "UI & Visuals",
+  "skyblocker.category.helper": "Helper",
+  "skyblocker.category.farming": "Farming",
   "skyblocker.category.general": "General",
   "skyblocker.option.general.enableTips": "Enable Tips",
   "skyblocker.option.general.bars": "Health, Mana, Defence & XP Bars",
@@ -104,8 +107,8 @@
   "skyblocker.option.general.itemTooltip.enableAccessoriesHelper.@Tooltip[3]": "This accessory can be upgraded. Also tells you what tier of accessory you have in that family.",
   "skyblocker.option.general.itemTooltip.enableAccessoriesHelper.@Tooltip[4]": "You already own an accessory in the same family that is better than this one. Also tells you what tier of accessory you have in that family.",
   "skyblocker.option.general.itemTooltip.enableAccessoriesHelper.@Tooltip[5]": "You don't own any accessory from this family.",
-  "skyblocker.option.general.dungeonQuality": "Dungeon Quality",
-  "skyblocker.option.general.dungeonQuality.@Tooltip": "Displays quality and tier of dungeon drops from mobs.\n\n\nReminder:\nTier 1-3 dropped from F1-F3\nTier 4-7 dropped from F4-F7 or M1-M4\nTier 8-10 are dropped only from M5-M7",
+  "skyblocker.option.general.itemTooltip.dungeonQuality": "Dungeon Quality",
+  "skyblocker.option.general.itemTooltip.dungeonQuality.@Tooltip": "Displays quality and tier of dungeon drops from mobs.\n\n\nReminder:\nTier 1-3 dropped from F1-F3\nTier 4-7 dropped from F4-F7 or M1-M4\nTier 8-10 are dropped only from M5-M7",
   "skyblocker.option.general.enableNewYearCakesHelper": "Enable New Year Cakes Helper",
   "skyblocker.option.general.enableNewYearCakesHelper.@Tooltip": "Highlights the missing new year cakes green and the cakes you have already red.\n\nRequires you to open your cake bag at least once to work.",
   "skyblocker.option.general.itemInfoDisplay": "Item Info Display",
@@ -167,25 +170,26 @@
   "skyblocker.option.general.searchOverlay.enableCommands": "Open from commands",
   "skyblocker.option.general.searchOverlay.enableCommands.@Tooltip": "Opens the bazaar search with \"/bzs\" and auction house with \"/ahs\". A re-log is required for this setting to be updated.",
   "skyblocker.option.general.searchOverlay.historyLabel": "History:",
-  "skyblocker.option.general.betterPartyFinder": "Better Party Finder",
+  "skyblocker.option.general.fancyPartyFinder": "Fancy Party Finder",
   "skyblocker.option.general.fancyCraftingTable": "Fancy Crafting Table UI",
-  "skyblocker.option.general.betterAuctionHouse": "Better Auction House",
-  "skyblocker.option.general.betterAuctionHouse.enabled": "Enabled",
-  "skyblocker.option.general.betterAuctionHouse.highlightUnderAvgPrice": "Highlight BIN under average price",
+  "skyblocker.option.general.fancyAuctionHouse": "Fancy Auction House",
+  "skyblocker.option.general.fancyAuctionHouse.enabled": "Enabled",
+  "skyblocker.option.general.fancyAuctionHouse.highlightUnderAvgPrice": "Highlight BIN under average price",
   "skyblocker.itemTooltip.nullMessage": "§cItem price information on tooltip will renew in max 60 seconds. If not, check latest.log",
   "skyblocker.itemTooltip.noData": "§cNo Data",
 
   "skyblocker.wikiLookup.noArticleFound": "§rUnable to locate a wiki article for this item...",
 
-  "skyblocker.category.richPresence": "Discord Rich Presence",
-  "skyblocker.option.richPresence.info": "Skyblock Info",
-  "skyblocker.option.richPresence.info.PURSE": "PURSE",
-  "skyblocker.option.richPresence.info.BITS": "BITS",
-  "skyblocker.option.richPresence.info.LOCATION": "LOCATION",
-  "skyblocker.option.richPresence.info.@Tooltip": "This value doesn't matter if you are cycling",
-  "skyblocker.option.richPresence.cycleMode": "Cycle Skyblock Info",
-  "skyblocker.option.richPresence.enableRichPresence": "Enabled",
-  "skyblocker.option.richPresence.customMessage": "Custom Message",
+  "skyblocker.category.misc": "Misc",
+  "skyblocker.option.misc.richPresence": "Discord RPC",
+  "skyblocker.option.misc.richPresence.info": "Skyblock Info",
+  "skyblocker.option.misc.richPresence.info.PURSE": "PURSE",
+  "skyblocker.option.misc.richPresence.info.BITS": "BITS",
+  "skyblocker.option.misc.richPresence.info.LOCATION": "LOCATION",
+  "skyblocker.option.misc.richPresence.info.@Tooltip": "This value doesn't matter if you are cycling",
+  "skyblocker.option.misc.richPresence.cycleMode": "Cycle Skyblock Info",
+  "skyblocker.option.misc.richPresence.enableRichPresence": "Enabled",
+  "skyblocker.option.misc.richPresence.customMessage": "Custom Message",
 
   "skyblocker.category.quickNav": "Quick Navigation",
   "skyblocker.option.quickNav.enableQuickNav": "Enable Quick Navigation",
@@ -202,7 +206,7 @@
   "skyblocker.option.general.itemList": "Item List",
   "skyblocker.option.general.itemList.enableItemList": "Enable Item List",
 
-  "skyblocker.category.locations": "Locations",
+  "skyblocker.category.locations": "Other Locations",
   "skyblocker.option.locations.barn": "Barn",
   "skyblocker.option.locations.barn.solveHungryHiker": "Solve Hungry Hiker",
   "skyblocker.option.locations.barn.solveTreasureHunter": "Solve Treasure Hunter",
@@ -318,7 +322,10 @@
   "skyblocker.option.locations.dungeons.mimicMessage.sendMimicMessage.@Tooltip": "Sends a message in chat upon killing a mimic for other players' score calculation mods.",
   "skyblocker.option.locations.dungeons.mimicMessage.mimicMessage": "Mimic Message",
   "skyblocker.option.locations.dungeons.mimicMessage.mimicMessage.@Tooltip": "Message which will be sent in the chat upon killing a mimic. Recommended to retain the default value.",
-  "skyblocker.option.locations.dungeons.lividColor": "Livid Color",
+  "skyblocker.option.locations.dungeons.livid": "Livid (F5/M5)",
+  "skyblocker.option.locations.dungeons.map": "Map",
+  "skyblocker.option.locations.dungeons.puzzle": "Puzzle Solver",
+  "skyblocker.option.locations.dungeons.professor": "The Professor (F3/M3)",
   "skyblocker.option.locations.dungeons.lividColor.enableLividColorGlow": "Enable Livid Color Glow",
   "skyblocker.option.locations.dungeons.lividColor.enableLividColorGlow.@Tooltip": "Applies the glowing effect to the correct Livid in F5/M5.",
   "skyblocker.option.locations.dungeons.lividColor.enableLividColorText": "Enable Livid Color Text",
@@ -327,7 +334,7 @@
   "skyblocker.option.locations.dungeons.lividColor.enableLividColorTitle.@Tooltip": "Display the livid color in the title during the Livid boss fight.",
   "skyblocker.option.locations.dungeons.lividColor.lividColorText": "Livid Color Text",
   "skyblocker.option.locations.dungeons.lividColor.lividColorText.@Tooltip": "Text which will be sent in the chat during the Livid boss fight. The string \"[color]\" will be replaced with the livid color.",
-  "skyblocker.option.locations.dungeons.terminals": "Terminal Solvers",
+  "skyblocker.option.locations.dungeons.terminals": "Terminal Solvers (F7/M7)",
   "skyblocker.option.locations.dungeons.terminals.solveColor": "Solve Select Colored",
   "skyblocker.option.locations.dungeons.terminals.solveOrder": "Solve Click In Order",
   "skyblocker.option.locations.dungeons.terminals.solveStartsWith": "Solve Starts With",
@@ -338,7 +345,7 @@
   "skyblocker.option.locations.dungeons.allowDroppingProtectedItems": "Enabled Dropping Protected Items & on Locked Slots",
   "skyblocker.option.locations.dungeons.allowDroppingProtectedItems.@Tooltip": "Allows the use of class abilities in Dungeons on a locked slot or while holding an item which has been protected using /skyblocker protectItem.",
 
-  "skyblocker.option.locations.dwarvenMines": "Dwarven Mines",
+  "skyblocker.option.locations.dwarvenMines": "Mining",
   "skyblocker.option.locations.dwarvenMines.enableDrillFuel": "Enable Drill Fuel",
   "skyblocker.option.locations.dwarvenMines.solveFetchur": "Solve Fetchur",
   "skyblocker.option.locations.dwarvenMines.solvePuzzler": "Solve Puzzler Puzzle",
