+++ src/main/resources/assets/skyblocker/lang/es_es.json
@@ -21,14 +21,14 @@
     "skyblocker.option.general.hitbox.oldFarmlandHitbox": "Habilitar la Caja de Colisión de la 1.8 para la tierra de cultivo",
     "skyblocker.option.general.hitbox.oldLeverHitbox": "Habilitar la Caja de Colisión de la 1.8 para la palanca",
     "skyblocker.itemTooltip.noData": "§cSin Información",
-    "skyblocker.category.richPresence": "Discord Rich Presence",
-    "skyblocker.option.richPresence.info": "Información de Skyblock",
-    "skyblocker.option.richPresence.info.PURSE": "CARTERA",
-    "skyblocker.option.richPresence.info.BITS": "BITS",
-    "skyblocker.option.richPresence.info.LOCATION": "LOCALIZACIÓN",
-    "skyblocker.option.richPresence.cycleMode": "Ciclar Información de Skyblock",
-    "skyblocker.option.richPresence.enableRichPresence": "Habilitado",
-    "skyblocker.option.richPresence.customMessage": "Mensaje Personalizado",
+    "skyblocker.option.misc.richPresence": "Discord Rich Presence",
+    "skyblocker.option.misc.richPresence.info": "Información de Skyblock",
+    "skyblocker.option.misc.richPresence.info.PURSE": "CARTERA",
+    "skyblocker.option.misc.richPresence.info.BITS": "BITS",
+    "skyblocker.option.misc.richPresence.info.LOCATION": "LOCALIZACIÓN",
+    "skyblocker.option.misc.richPresence.cycleMode": "Ciclar Información de Skyblock",
+    "skyblocker.option.misc.richPresence.enableRichPresence": "Habilitado",
+    "skyblocker.option.misc.richPresence.customMessage": "Mensaje Personalizado",
     "skyblocker.category.quickNav": "Navegación Rápida",
     "skyblocker.option.quickNav.enableQuickNav": "Habilitar Navegación Rápida",
     "skyblocker.option.messages.hideImplosion": "Ocultar el mensaje de Implosion",
@@ -36,7 +36,7 @@
     "skyblocker.option.locations.dungeons.croesusHelper.@Tooltip": "Obscurece los cofres que ya han sido abiertos.",
     "skyblocker.option.general.itemTooltip": "Información extra de los objetos",
     "skyblocker.itemTooltip.nullMessage": "§cEl precio en la información en los objetos se actualiza cada 60 segundos. De lo contrario revisa lastest.log",
-    "skyblocker.option.richPresence.info.@Tooltip": "Este valor no importa si estas ciclando",
+    "skyblocker.option.misc.richPresence.info.@Tooltip": "Este valor no importa si estas ciclando",
     "skyblocker.option.messages.hideAds": "Ocultar anuncios del Chat Publico",
     "skyblocker.option.messages.hideTeleportPad": "Ocultar mensajes del Pad de Teletransporte",
     "skyblocker.option.messages.hideAutopet": "Ocultar mensajes del Autopet",
