+++ src/main/resources/assets/skyblocker/lang/fr_fr.json
@@ -20,12 +20,12 @@
     "skyblocker.option.general.hitbox.oldLeverHitbox": "Hitbox 1.8 du levier",
     "skyblocker.itemTooltip.nullMessage": "§cLes informations sur le prix des objets vont s'actualiser dans 60 secondes maximum. Sinon consultez le fichier latest.log",
     "skyblocker.itemTooltip.noData": "§cPas de données",
-    "skyblocker.category.richPresence": "Rich Presence Discord",
-    "skyblocker.option.richPresence.info": "Information à afficher",
-    "skyblocker.option.richPresence.info.@Tooltip": "Cette valeur n'a aucune importance si vous faites défiler",
-    "skyblocker.option.richPresence.cycleMode": "Faire défiler les informations",
-    "skyblocker.option.richPresence.enableRichPresence": "Activée",
-    "skyblocker.option.richPresence.customMessage": "Message personnalisé",
+    "skyblocker.option.misc.richPresence": "Rich Presence Discord",
+    "skyblocker.option.misc.richPresence.info": "Information à afficher",
+    "skyblocker.option.misc.richPresence.info.@Tooltip": "Cette valeur n'a aucune importance si vous faites défiler",
+    "skyblocker.option.misc.richPresence.cycleMode": "Faire défiler les informations",
+    "skyblocker.option.misc.richPresence.enableRichPresence": "Activée",
+    "skyblocker.option.misc.richPresence.customMessage": "Message personnalisé",
     "skyblocker.option.general.itemList": "Liste d'objets",
     "skyblocker.option.general.itemList.enableItemList": "Activer la liste d'objets",
     "skyblocker.category.locations": "Lieux",
@@ -59,7 +59,7 @@
     "skyblocker.option.messages.hideMana.@Tooltip": "Permet une meilleure expérience avec les barres",
     "skyblocker.option.locations.dungeons.mapScaling": "Taille de la Carte",
     "skyblocker.option.locations.dungeons.lividColor.lividColorText.@Tooltip": "Texte qui sera envoyé dans le chat lors du boss Livid. Le string \"[color]\" sera remplacé par la couleur Livid.",
-    "skyblocker.option.locations.dungeons.lividColor": "Couleur Livid",
+    "skyblocker.option.locations.dungeons.livid": "Couleur Livid",
     "skyblocker.option.locations.dungeons.lividColor.enableLividColorText": "Activer la Couleur Livid",
     "skyblocker.option.locations.dungeons.lividColor.enableLividColorText.@Tooltip": "Envoyer la Couler Livid dans le chat durant le boss Livid.",
     "skyblocker.option.locations.dungeons.lividColor.lividColorText": "Texte de la Couleur Livid",
@@ -80,7 +80,7 @@
     "skyblocker.option.general.tabHud": "HUD Tab joli (temporairement désactivé en dehors des donjons)",
     "skyblocker.option.general.tabHud.tabHudEnabled": "Activer le HUD Tab joli",
     "skyblocker.option.general.tabHud.tabHudScale": "Taille du HUD Tab joli",
-    "skyblocker.option.richPresence.info.LOCATION": "LIEU",
+    "skyblocker.option.misc.richPresence.info.LOCATION": "LIEU",
     "skyblocker.category.quickNav": "Navigation rapide",
     "skyblocker.option.quickNav.enableQuickNav": "Activer la navigation rapide",
     "skyblocker.updateRepository.error": "§cLa mise à jour de la repository a échoué. Supprimez les fichiers manuellement et relancez le jeu.",
@@ -89,8 +89,8 @@
     "skyblocker.option.general.fishing": "Assistant de pêche",
     "skyblocker.option.general.itemTooltip.avg.THREE_DAY": "Prix pour 3 jours",
     "skyblocker.option.general.itemTooltip.avg.BOTH": "Les deux",
-    "skyblocker.option.richPresence.info.PURSE": "PORTE-MONNAIE",
-    "skyblocker.option.richPresence.info.BITS": "BITS",
+    "skyblocker.option.misc.richPresence.info.PURSE": "PORTE-MONNAIE",
+    "skyblocker.option.misc.richPresence.info.BITS": "BITS",
     "skyblocker.option.general.experiments": "Aide aux expériences de la table d'expérimentation",
     "skyblocker.option.general.experiments.enableChronomatronSolver": "Activer l'aide au Chronomatron",
     "skyblocker.option.general.experiments.enableSuperpairsSolver": "Activer l'aide aux Superpairs",
@@ -162,8 +162,8 @@
     "skyblocker.option.general.itemInfoDisplay.attributeShardInfo.@Tooltip": "Afficher le niveau de l'attribut avec le nombre d'items et les initiales du nom de l'attribut.",
     "skyblocker.option.general.wikiLookup.officialWiki": "Utiliser le wiki officiel",
     "skyblocker.option.general.itemProtection.slotLockStyle": "Style de l'icône de verrouillage d'emplacement",
-    "skyblocker.option.general.dungeonQuality": "Qualité du donjon",
-    "skyblocker.option.general.dungeonQuality.@Tooltip": "Affiche la qualité et le niveau des objets lâchés par les ennemis des donjons à leur mort.\n\n\nRappel:\nNiveau 1-3 lâchés en F1-F3\nNiveau 4-7 lâchés en F4-F7 ou M1-M4\nNiveau 8-10 lâchés en M5-M7",
+    "skyblocker.option.general.itemTooltip.dungeonQuality": "Qualité du donjon",
+    "skyblocker.option.general.itemTooltip.dungeonQuality.@Tooltip": "Affiche la qualité et le niveau des objets lâchés par les ennemis des donjons à leur mort.\n\n\nRappel:\nNiveau 1-3 lâchés en F1-F3\nNiveau 4-7 lâchés en F4-F7 ou M1-M4\nNiveau 8-10 lâchés en M5-M7",
     "skyblocker.option.general.itemInfoDisplay.attributeShardInfo": "Info des fragments d'attributs",
     "skyblocker.option.general.specialEffects.rareDungeonDropEffects": "Effet de drop de donjon rare",
     "skyblocker.option.general.specialEffects.rareDungeonDropEffects.@Tooltip": "Ajoute un effet visuel spécial activé en obtenant un objet de donjon rare!",
@@ -194,7 +194,7 @@
     "skyblocker.option.locations.dwarvenMines.crystalsHud.enabled": "Activé",
     "skyblocker.option.locations.dwarvenMines.dwarvenHud.screen": "Configuration du HUD des Nains...",
     "skyblocker.option.locations.dwarvenMines.crystalsHud": "Carte de Crystal Hollows",
-    "skyblocker.option.general.betterPartyFinder": "Recherche de Groupe Amélioré",
+    "skyblocker.option.general.fancyPartyFinder": "Recherche de Groupe Amélioré",
     "skyblocker.wikiLookup.noArticleFound": "§rImpossivle de trouver un article sur le wiki pour cet item...",
     "skyblocker.option.quickNav.button.clickEvent": "Événement de clique",
     "skyblocker.option.general.flameOverlay.flameHeight.@Tooltip": "100%  taille par défaut\n0%     désactivé",
@@ -237,9 +237,9 @@
     "skyblocker.option.general.searchOverlay.enableCommands.@Tooltip": "Ouvrir la recherche du bazaar avec \"/bzs\" et l'auction house avec \"/ahs\". Une reconnection sera nécessaire pour que ce paramètre prenne effet.",
     "skyblocker.option.general.searchOverlay.historyLabel": "Historique :",
     "skyblocker.option.general.fancyCraftingTable": "UI de Table de Craft joli",
-    "skyblocker.option.general.betterAuctionHouse": "Auction House Amélioré",
-    "skyblocker.option.general.betterAuctionHouse.enabled": "Activé",
-    "skyblocker.option.general.betterAuctionHouse.highlightUnderAvgPrice": "Surlignzr les BIN est dessous du prix moyen",
+    "skyblocker.option.general.fancyAuctionHouse": "Auction House Amélioré",
+    "skyblocker.option.general.fancyAuctionHouse.enabled": "Activé",
+    "skyblocker.option.general.fancyAuctionHouse.highlightUnderAvgPrice": "Surlignzr les BIN est dessous du prix moyen",
     "skyblocker.option.locations.crimsonIsle.kuudra": "Kuudra",
     "skyblocker.option.locations.crimsonIsle.kuudra.safeSpotWaypoints": "Waypoints des endroits sûrs",
     "skyblocker.option.locations.crimsonIsle.kuudra.noArrowPoisonWarning": "Avertissement pas de flèches empoisonnées",
