+++ src/main/resources/assets/skyblocker/lang/id_id.json
@@ -21,8 +21,8 @@
     "skyblocker.option.general.itemTooltip.enableAvgBIN": "Aktifkan Rata-rata Harga BIN",
     "skyblocker.option.general.itemTooltip.avg": "Tipe Rata-rata",
     "skyblocker.option.general.itemTooltip.enableBazaarPrice": "Aktifkan Harga Jual Beli Bazaar",
-    "skyblocker.option.richPresence.info": "Informasi Skyblock",
-    "skyblocker.option.richPresence.enableRichPresence": "Diaktifkan",
+    "skyblocker.option.misc.richPresence.info": "Informasi Skyblock",
+    "skyblocker.option.misc.richPresence.enableRichPresence": "Diaktifkan",
     "skyblocker.option.general.itemList": "Daftar Item",
     "skyblocker.option.general.itemList.enableItemList": "Aktifkan Daftar Item",
     "key.wikiLookup": "Pencarian Wiki",
