+++ src/main/resources/assets/skyblocker/lang/it_it.json
@@ -40,16 +40,16 @@
     "skyblocker.option.general.quiverWarning.enableQuiverWarningInDungeons": "Attiva avviso per faretra nei Dungeons",
     "skyblocker.option.general.itemTooltip": "Descrizione Oggetto",
     "skyblocker.option.general.itemTooltip.avg.BOTH": "Entrambi",
-    "skyblocker.option.general.dungeonQuality": "Qualita Dungeon",
+    "skyblocker.option.general.itemTooltip.dungeonQuality": "Qualita Dungeon",
     "skyblocker.option.general.wikiLookup": "Guarda sulla Wiki",
     "skyblocker.option.general.chestValue": "Valore Cesta",
     "skyblocker.option.general.specialEffects": "Effetti Speciali",
     "skyblocker.option.general.hitbox": "Hitbox",
     "skyblocker.option.general.teleportOverlay": "Overlay per Teletrasporto",
     "skyblocker.option.general.searchOverlay.historyLabel": "Storia:",
-    "skyblocker.option.richPresence.info.PURSE": "Portafoglio",
-    "skyblocker.option.richPresence.info.BITS": "BIT",
-    "skyblocker.option.richPresence.info.LOCATION": "POSIZIONE",
+    "skyblocker.option.misc.richPresence.info.PURSE": "Portafoglio",
+    "skyblocker.option.misc.richPresence.info.BITS": "BIT",
+    "skyblocker.option.misc.richPresence.info.LOCATION": "POSIZIONE",
     "skyblocker.option.quickNav.button.item": "Oggetto",
     "skyblocker.option.quickNav.button.item.nbt": "NBT",
     "skyblocker.option.locations.crimsonIsle.kuudra": "Kuudra",
@@ -92,7 +92,7 @@
     "skyblocker.option.general.itemTooltip.enableAccessoriesHelper.@Tooltip[2]": "Questo accessorio è un miglioramento rispetto a quello della stessa famiglia che già possiedi. Mostra anche di che livello è l'accessorio all'interno della sua famiglia.",
     "skyblocker.option.general.itemTooltip.enableAccessoriesHelper.@Tooltip[0]": "Quando passi il mouse sopra un accessorio vieni informato se ce lo hai gia, o meno, e se è meglio o peggio di quello che hai già. Lista dei possibili status:",
     "skyblocker.option.general.itemTooltip.enableMuseumInfo.@Tooltip": "SI questo oggetto è donabile al museo, viene mostrata la categoria del museo a cui appartiene. Mostra anche se hai già donato o meno quel'oggetto.\n\nFai attenzione ad aver abilitato l'API per il Museo per far si di avere informazioni accurate!",
-    "skyblocker.option.richPresence.enableRichPresence": "Attivo",
+    "skyblocker.option.misc.richPresence.enableRichPresence": "Attivo",
     "skyblocker.option.quickNav.button.render": "Rendering",
     "skyblocker.category.locations": "Luoghi",
     "skyblocker.option.locations.barn": "Barn",
@@ -115,8 +115,8 @@
     "skyblocker.option.general.flameOverlay.flameHeight": "Altezza della Fiamma",
     "skyblocker.option.general.flameOverlay.flameOpacity": "Opacità della FIamma",
     "skyblocker.itemTooltip.noData": "§cNessuna Informazione",
-    "skyblocker.option.richPresence.info": "Informazioni Skyblock",
-    "skyblocker.option.richPresence.customMessage": "Messaggio Personalizzato",
+    "skyblocker.option.misc.richPresence.info": "Informazioni Skyblock",
+    "skyblocker.option.misc.richPresence.customMessage": "Messaggio Personalizzato",
     "skyblocker.category.quickNav": "Navigazione Veloce",
     "skyblocker.option.quickNav.button": "Bottone %d",
     "skyblocker.option.quickNav.button.item.itemName": "ID Oggetto",
@@ -138,7 +138,7 @@
     "skyblocker.option.locations.dungeons.dungeonChestProfit.lossColor": "Colore per Perdita",
     "skyblocker.option.locations.dungeons.mapScaling": "Dimensioni della Mappa",
     "skyblocker.option.locations.dungeons.mimicMessage": "Messaggio per il Mimic",
-    "skyblocker.option.locations.dungeons.lividColor": "Colore Livid",
+    "skyblocker.option.locations.dungeons.livid": "Colore Livid",
     "skyblocker.option.locations.dwarvenMines.solveFetchur": "Risolvi Fetchur",
     "skyblocker.option.locations.dwarvenMines.dwarvenHud": "HUD per le Dwarven Mines",
     "skyblocker.option.locations.dwarvenMines.crystalsHud.mapScaling": "Dimensioni Mappa",
@@ -176,7 +176,7 @@
     "skyblocker.option.locations.dungeons.dungeonChestProfit.enableProfitCalculator.@Tooltip": "Mostra il profitto di un baule del dungeon nel titolo dello schermo del baule.\nVerde se c'è profitto.\nRosso se non c'è profitto.\nGrigio se non guadagni o perdi nulla.\nBlu se i calcoli sono stati basati su dati incompleti.",
     "skyblocker.option.messages.chatRules.screen.ruleScreen.sounds": "Riproduci Suono:",
     "skyblocker.option.slayer.endermanSlayer": "Slayer di Enderman",
-    "skyblocker.option.general.dungeonQuality.@Tooltip": "Mostra la qualità e il livello dei drop del dungeon dai mob.\n\n\nPromemoria:\nLivello 1-3 droppato da F1-F3\nLivello 4-7 droppato da F4-F7 o M1-M4\n Livello 8-10 droppati solo da M5-M7",
+    "skyblocker.option.general.itemTooltip.dungeonQuality.@Tooltip": "Mostra la qualità e il livello dei drop del dungeon dai mob.\n\n\nPromemoria:\nLivello 1-3 droppato da F1-F3\nLivello 4-7 droppato da F4-F7 o M1-M4\n Livello 8-10 droppati solo da M5-M7",
     "skyblocker.option.locations.dungeons.dungeonScore.dungeonScoreMessage.@Tooltip": "Messaggio che verrà inviato in chat quando si raggiunge %d punteggio nei dungeon. La stringa “[punteggio]” sarà sostituita con il punteggio del dungeon (%d).",
     "skyblocker.option.general.itemTooltip.enableAccessoriesHelper.@Tooltip[4]": "Possiedi già un accessorio nella stessa famiglia che è migliore di questo. Dice anche di che livello è l’accessorio che hai in quella famiglia.",
     "skyblocker.option.general.enableNewYearCakesHelper.@Tooltip": "Evidenzia in verde le torte di capodanno mancanti e in rosso le torte che hai già. \n\nRichiede che tu apra almeno una volta la tua borsa delle torte per funzionare.",
