+++ src/main/resources/assets/skyblocker/lang/ja_jp.json
@@ -19,12 +19,12 @@
     "skyblocker.option.general.hitbox.oldFarmlandHitbox": "バージョン1.8での作物のヒットボックスを使う",
     "skyblocker.option.general.hitbox.oldLeverHitbox": "バージョン1.8でのレバーのヒットボックスを使う",
     "skyblocker.itemTooltip.nullMessage": "§cツールチップ上のアイテムの値段は最大60秒ごとに更新されます。されていない場合はlatest.logを確認してください",
-    "skyblocker.category.richPresence": "discordに自分のゲームアクティビティを表示する",
-    "skyblocker.option.richPresence.info": "skyblockの情報",
-    "skyblocker.option.richPresence.info.@Tooltip": "この値はあなたがサイクリングしている場合は重要ではありません",
-    "skyblocker.option.richPresence.cycleMode": "円状のskyblockの情報",
-    "skyblocker.option.richPresence.enableRichPresence": "有効",
-    "skyblocker.option.richPresence.customMessage": "カスタムメッセージ",
+    "skyblocker.option.misc.richPresence": "discordに自分のゲームアクティビティを表示する",
+    "skyblocker.option.misc.richPresence.info": "skyblockの情報",
+    "skyblocker.option.misc.richPresence.info.@Tooltip": "この値はあなたがサイクリングしている場合は重要ではありません",
+    "skyblocker.option.misc.richPresence.cycleMode": "円状のskyblockの情報",
+    "skyblocker.option.misc.richPresence.enableRichPresence": "有効",
+    "skyblocker.option.misc.richPresence.customMessage": "カスタムメッセージ",
     "skyblocker.option.general.itemList": "アイテムリスト",
     "skyblocker.option.general.itemList.enableItemList": "アイテムリストを有効にする",
     "skyblocker.category.locations": "場所",
