+++ src/main/resources/assets/skyblocker/lang/ko_kr.json
@@ -23,15 +23,15 @@
     "skyblocker.option.general.hitbox.oldLeverHitbox": "1.8 레버 히트박스 활성화",
     "skyblocker.itemTooltip.nullMessage": "§c아이템 가격 툴팁이 최대 60초 이내에 갱신됩니다. 갱신되지 않을 시 latest.log 를 확인하세요.",
     "skyblocker.itemTooltip.noData": "§c데이터 없음",
-    "skyblocker.category.richPresence": "디스코드 Rich Presence",
-    "skyblocker.option.richPresence.info": "스카이블록 정보",
-    "skyblocker.option.richPresence.info.PURSE": "PURSE",
-    "skyblocker.option.richPresence.info.BITS": "BITS",
-    "skyblocker.option.richPresence.info.LOCATION": "위치",
-    "skyblocker.option.richPresence.info.@Tooltip": "이 값은 순환중일 시 중요하지 않습니다",
-    "skyblocker.option.richPresence.cycleMode": "스카이블록 정보 순환",
-    "skyblocker.option.richPresence.enableRichPresence": "활성화됨",
-    "skyblocker.option.richPresence.customMessage": "사용자 지정 메시지",
+    "skyblocker.option.misc.richPresence": "디스코드 Rich Presence",
+    "skyblocker.option.misc.richPresence.info": "스카이블록 정보",
+    "skyblocker.option.misc.richPresence.info.PURSE": "PURSE",
+    "skyblocker.option.misc.richPresence.info.BITS": "BITS",
+    "skyblocker.option.misc.richPresence.info.LOCATION": "위치",
+    "skyblocker.option.misc.richPresence.info.@Tooltip": "이 값은 순환중일 시 중요하지 않습니다",
+    "skyblocker.option.misc.richPresence.cycleMode": "스카이블록 정보 순환",
+    "skyblocker.option.misc.richPresence.enableRichPresence": "활성화됨",
+    "skyblocker.option.misc.richPresence.customMessage": "사용자 지정 메시지",
     "skyblocker.category.quickNav": "즐겨찾기",
     "skyblocker.option.quickNav.enableQuickNav": "즐겨찾기 활성화",
     "skyblocker.option.general.itemList": "아이템 목록",
