+++ src/main/resources/assets/skyblocker/lang/nb_no.json
@@ -17,15 +17,15 @@
     "skyblocker.option.general.hitbox.oldFarmlandHitbox": "Aktiver 1.8 treffboks for avlinger",
     "skyblocker.option.general.hitbox.oldLeverHitbox": "Aktiver 1.8 spak Treffboks",
     "skyblocker.itemTooltip.noData": "§cIngen data",
-    "skyblocker.category.richPresence": "Discord-rik tilstedeværelse",
-    "skyblocker.option.richPresence.info": "Skyblock-informasjon",
-    "skyblocker.option.richPresence.info.PURSE": "VESKE",
-    "skyblocker.option.richPresence.info.BITS": "BITS",
-    "skyblocker.option.richPresence.info.LOCATION": "PLASSERING",
-    "skyblocker.option.richPresence.info.@Tooltip": "Denne verdien har ikke noe å si hvis du cycling",
-    "skyblocker.option.richPresence.cycleMode": "Cycle Skyblock Informasjon",
-    "skyblocker.option.richPresence.enableRichPresence": "Aktivert",
-    "skyblocker.option.richPresence.customMessage": "Egendefinert melding",
+    "skyblocker.option.misc.richPresence": "Discord-rik tilstedeværelse",
+    "skyblocker.option.misc.richPresence.info": "Skyblock-informasjon",
+    "skyblocker.option.misc.richPresence.info.PURSE": "VESKE",
+    "skyblocker.option.misc.richPresence.info.BITS": "BITS",
+    "skyblocker.option.misc.richPresence.info.LOCATION": "PLASSERING",
+    "skyblocker.option.misc.richPresence.info.@Tooltip": "Denne verdien har ikke noe å si hvis du cycling",
+    "skyblocker.option.misc.richPresence.cycleMode": "Cycle Skyblock Informasjon",
+    "skyblocker.option.misc.richPresence.enableRichPresence": "Aktivert",
+    "skyblocker.option.misc.richPresence.customMessage": "Egendefinert melding",
     "skyblocker.category.quickNav": "Rask navigering",
     "skyblocker.option.quickNav.enableQuickNav": "Aktiver hurtignavigering",
     "key.categories.skyblocker": "SkyBlocker",
