+++ src/main/resources/assets/skyblocker/lang/pt_br.json
@@ -23,12 +23,12 @@
     "skyblocker.option.general.teleportOverlay.enableSinrecallTransmission": "Ativar Overlay de Transmissão de Sinrecall",
     "skyblocker.option.general.teleportOverlay.enableWitherImpact": "Ativar Overlay de Impacto de Wither",
     "skyblocker.itemTooltip.noData": "§cSem Dados",
-    "skyblocker.option.richPresence.info": "Informação do Skyblock",
-    "skyblocker.option.richPresence.info.PURSE": "BOLSA",
-    "skyblocker.option.richPresence.info.@Tooltip": "Esse valor não importa se você está passando por um ciclo",
-    "skyblocker.option.richPresence.cycleMode": "Informação de Ciclo do Skyblock",
-    "skyblocker.option.richPresence.enableRichPresence": "Ativado",
-    "skyblocker.option.richPresence.customMessage": "Mensagem Personalizada",
+    "skyblocker.option.misc.richPresence.info": "Informação do Skyblock",
+    "skyblocker.option.misc.richPresence.info.PURSE": "BOLSA",
+    "skyblocker.option.misc.richPresence.info.@Tooltip": "Esse valor não importa se você está passando por um ciclo",
+    "skyblocker.option.misc.richPresence.cycleMode": "Informação de Ciclo do Skyblock",
+    "skyblocker.option.misc.richPresence.enableRichPresence": "Ativado",
+    "skyblocker.option.misc.richPresence.customMessage": "Mensagem Personalizada",
     "skyblocker.option.quickNav.enableQuickNav": "Ativar Navegação Rápida",
     "key.hotbarSlotLock": "Trancar Slot (Hotbar)",
     "skyblocker.title": "Configurações do Skyblocker",
@@ -45,7 +45,7 @@
     "skyblocker.option.general.teleportOverlay.enableTeleportOverlays": "Ativar Overlays de Teleporte",
     "skyblocker.option.general.teleportOverlay.enableEtherTransmission": "Ativar Overlay de Transmissão Éter",
     "skyblocker.itemTooltip.nullMessage": "§cA Informação do preço do item no tooltip irá se renovar no máximo 60 segundos. Caso contrário, cheque \"latest.log\"",
-    "skyblocker.option.richPresence.info.LOCATION": "LOCALIZAÇÃO",
+    "skyblocker.option.misc.richPresence.info.LOCATION": "LOCALIZAÇÃO",
     "skyblocker.category.quickNav": "Navegação Rápida",
     "key.categories.skyblocker": "Skyblocker",
     "skyblocker.option.general.etherwarpOverlay": "Overlay de Etherwarp",
@@ -66,8 +66,8 @@
     "skyblocker.option.general.itemTooltip.avg.@Tooltip": "Você pode escolher quantos dias de preço médio para ser",
     "skyblocker.option.general.itemInfoDisplay": "Exibição de Informação do Item",
     "skyblocker.option.general.itemInfoDisplay.attributeShardInfo": "Informação de Fragmento de Atributo",
-    "skyblocker.category.richPresence": "Discord Rich Presence",
-    "skyblocker.option.richPresence.info.BITS": "BITS",
+    "skyblocker.option.misc.richPresence": "Discord Rich Presence",
+    "skyblocker.option.misc.richPresence.info.BITS": "BITS",
     "skyblocker.category.messages": "Mensagens",
     "skyblocker.option.messages.hideImplosion": "Esconder Mensagem de Implosão",
     "skyblocker.option.messages.hideMoltenWave": "Esconder Mensagem do Molten Wave",
@@ -172,7 +172,7 @@
     "skyblocker.option.locations.dungeons.croesusHelper.@Tooltip": "Acinzentar os baús que já foram abertos.",
     "skyblocker.option.locations.dungeons.enableMap": "Ativar mapa",
     "skyblocker.option.locations.dungeons.mapScaling": "Tamanho do mapa",
-    "skyblocker.option.locations.dungeons.lividColor": "Cor do Livid",
+    "skyblocker.option.locations.dungeons.livid": "Cor do Livid",
     "skyblocker.option.locations.dungeons.lividColor.enableLividColorText": "Ativar cor para Livid",
     "skyblocker.option.locations.dungeons.terminals": "Guia dos terminais",
     "skyblocker.option.locations.dungeons.terminals.solveColor": "Guia de seleção colorido",
@@ -289,7 +289,7 @@
     "skyblocker.option.general.itemProtection.slotLockStyle.@Tooltip": "Escolhe entre ícone de slot bloqueado bonito ou classico.",
     "key.itemProtection": "Proteger Item",
     "skyblocker.option.general.chestValue.color": "Cor do Valor do Baú",
-    "skyblocker.option.general.dungeonQuality": "Qualidade da Dungeon",
+    "skyblocker.option.general.itemTooltip.dungeonQuality": "Qualidade da Dungeon",
     "skyblocker.option.general.tabHud.enableHudBackground": "Ativar Fundo de HUD",
     "skyblocker.option.general.tabHud.enableHudBackground.@Tooltip": "Ativa o fundo de uma HUD que não é a tab.",
     "skyblocker.option.general.waypoints": "Waypoints",
