+++ src/main/resources/assets/skyblocker/lang/ru_ru.json
@@ -33,12 +33,12 @@
     "skyblocker.option.general.itemTooltip.avg.@Tooltip": "Можно выбрать, за сколько дней будет браться средняя цена",
     "skyblocker.option.general.itemTooltip.enableLowestBIN": "Показать минимальную цену на BIN",
     "skyblocker.option.general.itemTooltip.enableBazaarPrice": "Показать цены на Bazaar-е",
-    "skyblocker.category.richPresence": "Активность в статусе Discord",
-    "skyblocker.option.richPresence.info": "Отображаемая информация",
-    "skyblocker.option.richPresence.info.@Tooltip": "Не имеет значения при цикличном отображении",
-    "skyblocker.option.richPresence.cycleMode": "Отображать все опции циклично",
-    "skyblocker.option.richPresence.enableRichPresence": "Включить",
-    "skyblocker.option.richPresence.customMessage": "Своё сообщение",
+    "skyblocker.option.misc.richPresence": "Активность в статусе Discord",
+    "skyblocker.option.misc.richPresence.info": "Отображаемая информация",
+    "skyblocker.option.misc.richPresence.info.@Tooltip": "Не имеет значения при цикличном отображении",
+    "skyblocker.option.misc.richPresence.cycleMode": "Отображать все опции циклично",
+    "skyblocker.option.misc.richPresence.enableRichPresence": "Включить",
+    "skyblocker.option.misc.richPresence.customMessage": "Своё сообщение",
     "skyblocker.option.general.itemList": "Список предметов",
     "skyblocker.option.general.itemList.enableItemList": "Включить Список Предметов",
     "skyblocker.option.messages.hideCombo": "Скрывать сообщения о комбо",
@@ -60,8 +60,8 @@
     "skyblocker.option.general.hitbox.oldLeverHitbox": "Использовать хитбокс рычагов из 1.8",
     "skyblocker.option.general.itemTooltip.avg.ONE_DAY": "Цена за 1 день",
     "skyblocker.itemTooltip.noData": "§cНет данных",
-    "skyblocker.option.richPresence.info.PURSE": "PURSE",
-    "skyblocker.option.richPresence.info.LOCATION": "LOCATION",
+    "skyblocker.option.misc.richPresence.info.PURSE": "PURSE",
+    "skyblocker.option.misc.richPresence.info.LOCATION": "LOCATION",
     "skyblocker.category.quickNav": "Быстрый Доступ",
     "skyblocker.option.quickNav.enableQuickNav": "Включить кнопки в инвентаре",
     "skyblocker.option.general.tabHud": "Красивое TAB меню(временно отключено вне данжа)",
@@ -74,7 +74,7 @@
     "skyblocker.option.locations.dwarvenMines.dwarvenHud.style.@Tooltip[1]": "\nКрасивый: Показывает название, процент и шкалу выполнения, а также иконку.",
     "skyblocker.option.locations.dungeons.mapScaling": "Размер Карты",
     "skyblocker.option.locations.dungeons.lividColor.enableLividColorText": "Включить Цвет Босса Livid",
-    "skyblocker.option.locations.dungeons.lividColor": "Цвет Босса Livid",
+    "skyblocker.option.locations.dungeons.livid": "Цвет Босса Livid",
     "skyblocker.option.locations.dungeons.lividColor.enableLividColorText.@Tooltip": "Отправляет в чат информацию о том, какого цвета босс Livid.",
     "skyblocker.option.locations.dungeons.lividColor.lividColorText": "Текст О Цвете Livid",
     "skyblocker.option.locations.dwarvenMines.dwarvenHud.style": "Стиль HUD",
@@ -89,7 +89,7 @@
     "skyblocker.option.messages.chatFilterResult.FILTER": "Скрыть",
     "skyblocker.option.messages.chatFilterResult.PASS": "Не скрывать",
     "skyblocker.updateRepository.error": "§cОшибка в обновлении местного репозитория. Перезапустите игру, удалив файлы вручную.",
-    "skyblocker.option.richPresence.info.BITS": "BITS",
+    "skyblocker.option.misc.richPresence.info.BITS": "BITS",
     "skyblocker.option.locations.dungeons.croesusHelper": "Помощь в меню Croesus",
     "skyblocker.option.general.experiments": "Помощь в Экспериментах",
     "skyblocker.option.general.experiments.enableUltrasequencerSolver": "Помощь с Ultrasequencer",
@@ -197,7 +197,7 @@
     "skyblocker.option.locations.dungeons.secretWaypoints.enableRoomMatching": "Включить нумерацию комнат",
     "skyblocker.option.locations.dungeons.secretWaypoints.enableDefaultWaypoints": "Включить стандартные метки",
     "skyblocker.wikiLookup.noArticleFound": "§rДля этого предмета не найдено статьи на Вики...",
-    "skyblocker.option.general.betterPartyFinder": "Улучшенный поиск компании",
+    "skyblocker.option.general.fancyPartyFinder": "Улучшенный поиск компании",
     "skyblocker.option.locations.spidersDen.relics.highlightFoundRelics": "Подсветка найденных реликвий",
     "skyblocker.option.locations.spidersDen.relics.enableRelicsHelper": "Включить помощника по хайд-реликвиям",
     "skyblocker.option.locations.dungeons.secretWaypoints.showSecretText": "Включить секретный текст",
@@ -321,8 +321,8 @@
     "skyblocker.option.general.itemProtection": "Предметная защита",
     "skyblocker.option.general.itemProtection.slotLockStyle": "Стиль слота с закреплённым предметом",
     "skyblocker.option.locations.crimsonIsle.kuudra": "Kuudrа",
-    "skyblocker.option.general.dungeonQuality": "Данж-качество",
-    "skyblocker.option.general.dungeonQuality.@Tooltip": "Показывает качество и уровень дропа из данжа с мобов.\n\n\nПример:\nУр. 1-3 падает на 1-3 этажах\nУр. 4-7 падает на 4-7 этажах, или с 1-4 ур. Мастера\nУр. 8-10 падает только с 5-7 ур. Мастера",
+    "skyblocker.option.general.itemTooltip.dungeonQuality": "Данж-качество",
+    "skyblocker.option.general.itemTooltip.dungeonQuality.@Tooltip": "Показывает качество и уровень дропа из данжа с мобов.\n\n\nПример:\nУр. 1-3 падает на 1-3 этажах\nУр. 4-7 падает на 4-7 этажах, или с 1-4 ур. Мастера\nУр. 8-10 падает только с 5-7 ур. Мастера",
     "skyblocker.option.locations.dungeons.dungeonChestProfit.croesusProfit": "Подсчёт окупаемости сундука Croesus",
     "skyblocker.option.locations.dungeons.dungeonChestProfit.croesusProfit.@Tooltip": "Подсвечивает сундук с самой лучшей окупаемостью у Croesus НИПа.\n\n\nСамый лучший сундук будет подсвечиваться зелёным цветом.\nЕсли для этого сундука нужен ключ данжа, то он будет подсвечиваться жёлтым",
     "skyblocker.option.locations.dungeons.mapScreen": "Расположение карты и результата забега...",
@@ -373,9 +373,9 @@
     "skyblocker.option.general.fishing.hideOtherPlayers.@Tooltip": "Не дает отображаться удочкам других игроков",
     "skyblocker.option.general.itemTooltip.enableAccessoriesHelper.@Tooltip[4]": "У вас уже есть аксессуар лучше, чем этот. Также сообщает, какого уровня имеющийся аксессуар.",
     "skyblocker.option.general.itemTooltip.enableAccessoriesHelper.@Tooltip[5]": "У вас нет аксессуаров из этого семейства.",
-    "skyblocker.option.general.betterAuctionHouse": "Улучшенный аукцион",
-    "skyblocker.option.general.betterAuctionHouse.enabled": "Включить",
-    "skyblocker.option.general.betterAuctionHouse.highlightUnderAvgPrice": "Подсвечивать BIN-аукционы по цене ниже средней",
+    "skyblocker.option.general.fancyAuctionHouse": "Улучшенный аукцион",
+    "skyblocker.option.general.fancyAuctionHouse.enabled": "Включить",
+    "skyblocker.option.general.fancyAuctionHouse.highlightUnderAvgPrice": "Подсвечивать BIN-аукционы по цене ниже средней",
     "skyblocker.option.locations.garden.farmingHud.enableHud": "Включить фермерский HUD",
     "skyblocker.option.locations.garden.farmingHud.config": "Настройка фермерского HUD...",
     "skyblocker.option.general.fishing.enableFishingTimer": "Включить Таймер",
