+++ src/main/resources/assets/skyblocker/lang/tr_tr.json
@@ -13,14 +13,14 @@
     "skyblocker.option.general.hitbox.oldFarmlandHitbox": "1.8 tarım hitboxlarını etkinleştir",
     "skyblocker.option.general.hitbox.oldLeverHitbox": "1.8 şalter hitboxunu etkinleştir",
     "skyblocker.itemTooltip.noData": "§cVeri yok",
-    "skyblocker.category.richPresence": "Discord Özel Durumu",
-    "skyblocker.option.richPresence.info": "Skyblock bilgisi",
-    "skyblocker.option.richPresence.info.PURSE": "Cüzdan",
-    "skyblocker.option.richPresence.info.BITS": "Bit",
-    "skyblocker.option.richPresence.info.LOCATION": "Konum",
-    "skyblocker.option.richPresence.info.@Tooltip": "Eğer döngüyü açtıysanız bu değer geçersiz kalır",
-    "skyblocker.option.richPresence.cycleMode": "Skyblock bilgilerini döngüye al",
-    "skyblocker.option.richPresence.customMessage": "Özel mesaj",
+    "skyblocker.option.misc.richPresence": "Discord Özel Durumu",
+    "skyblocker.option.misc.richPresence.info": "Skyblock bilgisi",
+    "skyblocker.option.misc.richPresence.info.PURSE": "Cüzdan",
+    "skyblocker.option.misc.richPresence.info.BITS": "Bit",
+    "skyblocker.option.misc.richPresence.info.LOCATION": "Konum",
+    "skyblocker.option.misc.richPresence.info.@Tooltip": "Eğer döngüyü açtıysanız bu değer geçersiz kalır",
+    "skyblocker.option.misc.richPresence.cycleMode": "Skyblock bilgilerini döngüye al",
+    "skyblocker.option.misc.richPresence.customMessage": "Özel mesaj",
     "skyblocker.option.locations.dwarvenMines.dwarvenHud": "Dwarven HUD",
     "skyblocker.option.locations.dwarvenMines.solveFetchur": "Fetchur'un mesajını çöz",
     "skyblocker.fishing.reelNow": "Şimdi sarıl!",
@@ -40,7 +40,7 @@
     "skyblocker.option.general.hideEmptyTooltips": "Menülerdeki boş eşya açıklamalarını gizle",
     "skyblocker.option.general.itemTooltip.avg.@Tooltip": "Kaç günlük ortalamanın gösterileceğini seçebilirsiniz",
     "skyblocker.option.general.itemTooltip.enableLowestBIN": "En düşük BIN fiyatını göster",
-    "skyblocker.option.richPresence.enableRichPresence": "Aktif",
+    "skyblocker.option.misc.richPresence.enableRichPresence": "Aktif",
     "skyblocker.option.locations.dwarvenMines.solvePuzzler": "Puzzler'ın bulmacasını çöz",
     "key.hotbarSlotLock": "Slot Kilidi (Araç Çubuğu)",
     "key.wikiLookup": "Wiki Araması",
