+++ src/main/resources/assets/skyblocker/lang/zh_cn.json
@@ -23,15 +23,15 @@
     "skyblocker.option.general.hitbox.oldLeverHitbox": "启用1.8的拉杆碰撞箱",
     "skyblocker.itemTooltip.nullMessage": "§c物品提示里的价格信息会在最多60秒内更新。如果没有，请检查latest.log",
     "skyblocker.itemTooltip.noData": "§c没有数据",
-    "skyblocker.category.richPresence": "Discord活动状态",
-    "skyblocker.option.richPresence.info": "Skyblock信息",
-    "skyblocker.option.richPresence.info.PURSE": "钱包",
-    "skyblocker.option.richPresence.info.BITS": "点数",
-    "skyblocker.option.richPresence.info.LOCATION": "位置",
-    "skyblocker.option.richPresence.info.@Tooltip": "如果您处于循环模式，这个值将不会生效",
-    "skyblocker.option.richPresence.cycleMode": "循环Skyblock信息",
-    "skyblocker.option.richPresence.enableRichPresence": "已启用",
-    "skyblocker.option.richPresence.customMessage": "自定义消息",
+    "skyblocker.option.misc.richPresence": "Discord活动状态",
+    "skyblocker.option.misc.richPresence.info": "Skyblock信息",
+    "skyblocker.option.misc.richPresence.info.PURSE": "钱包",
+    "skyblocker.option.misc.richPresence.info.BITS": "点数",
+    "skyblocker.option.misc.richPresence.info.LOCATION": "位置",
+    "skyblocker.option.misc.richPresence.info.@Tooltip": "如果您处于循环模式，这个值将不会生效",
+    "skyblocker.option.misc.richPresence.cycleMode": "循环Skyblock信息",
+    "skyblocker.option.misc.richPresence.enableRichPresence": "已启用",
+    "skyblocker.option.misc.richPresence.customMessage": "自定义消息",
     "skyblocker.category.quickNav": "快速导航",
     "skyblocker.option.quickNav.enableQuickNav": "启用快速导航",
     "skyblocker.option.general.itemList": "物品列表",
@@ -78,7 +78,7 @@
     "skyblocker.option.general.fishing.enableFishingHelper": "启用钓鱼助手",
     "skyblocker.fishing.reelNow": "收竿！",
     "skyblocker.option.locations.dungeons.lividColor.enableLividColorText": "启用真Livid的颜色提示文本",
-    "skyblocker.option.locations.dungeons.lividColor": "提示真Livid的颜色",
+    "skyblocker.option.locations.dungeons.livid": "提示真Livid的颜色",
     "skyblocker.option.locations.dungeons.lividColor.enableLividColorText.@Tooltip": "将真Livid的颜色发送到聊天栏。",
     "skyblocker.option.locations.dungeons.lividColor.lividColorText": "真Livid颜色提示信息文本",
     "skyblocker.option.locations.dungeons.lividColor.lividColorText.@Tooltip": "Livid Boss战时发送到聊天栏的信息， 字段 “[color]” 将被替换为真Livid的颜色。",
@@ -170,7 +170,7 @@
     "skyblocker.wikiLookup.noArticleFound": "§r无法找到此物品的维基条目...",
     "skyblocker.option.locations.dungeons.secretWaypoints.enableWitherWaypoints": "启用凋灵精粹路径点",
     "skyblocker.option.locations.dungeons.secretWaypoints.enableBatWaypoints": "启用蝙蝠路径点",
-    "skyblocker.option.general.betterPartyFinder": "更好的组队查找",
+    "skyblocker.option.general.fancyPartyFinder": "更好的组队查找",
     "skyblocker.option.locations.dungeons.secretWaypoints.enableRoomMatching.@Tooltip": "关闭此选项可节省约20MB左右的内存，但是秘密路径点和§l部分谜题助手功能§r需要启用该选项。",
     "skyblocker.option.locations.dungeons.secretWaypoints.enableItemWaypoints": "启用物品路径点",
     "skyblocker.option.general.quiverWarning.enableQuiverWarningAfterDungeon": "在地牢结束后启用箭袋提示",
@@ -192,7 +192,7 @@
     "skyblocker.option.general.specialEffects": "特效",
     "skyblocker.option.locations.dungeons.doorHighlight.doorHighlightType.secretWaypointsNote": "\n\n\n注意：若使此功能生效，必须启用地牢秘密路径点。",
     "skyblocker.option.locations.dungeons.dungeonScore.enableScoreHUD.@Tooltip": "在状态栏显示地牢的分数。",
-    "skyblocker.option.general.dungeonQuality": "地牢物品品质",
+    "skyblocker.option.general.itemTooltip.dungeonQuality": "地牢物品品质",
     "skyblocker.option.locations.dungeons.secretWaypoints.enableStonkWaypoints": "启用Stonk路径点",
     "skyblocker.option.messages.hideToggleSkyMall": "隐藏Sky Mall切换的消息",
     "skyblocker.option.locations.dwarvenMines.dwarvenHud.screen": "矮人矿井HUD配置",
@@ -528,7 +528,7 @@
     "text.skyblocker.discord": "Discord",
     "text.skyblocker.source": "源代码",
     "text.skyblocker.confirm": "确认",
-    "skyblocker.option.general.betterAuctionHouse.enabled": "启用",
+    "skyblocker.option.general.fancyAuctionHouse.enabled": "启用",
     "skyblocker.fancyAuctionHouse.editBid": "点击编辑出价！",
     "skyblocker.fancyAuctionHouse.alreadyTopBid": "已经是最高出价！",
     "skyblocker.fancyAuctionHouse.didntWin": "没有赢得拍卖 :(",
@@ -539,8 +539,8 @@
     "skyblocker.fancyAuctionHouse.yourAuction": "这是你的拍卖！",
     "skyblocker.fancyAuctionHouse.youPay": "你支付：%s",
     "skyblocker.fancyAuctionHouse.collectBid": "收集出价",
-    "skyblocker.option.general.betterAuctionHouse": "更好的拍卖行界面",
-    "skyblocker.option.general.betterAuctionHouse.highlightUnderAvgPrice": "高亮显示低于平均价格的BIN物品",
+    "skyblocker.option.general.fancyAuctionHouse": "更好的拍卖行界面",
+    "skyblocker.option.general.fancyAuctionHouse.highlightUnderAvgPrice": "高亮显示低于平均价格的BIN物品",
     "skyblocker.fancyAuctionHouse.price": "价格：",
     "skyblocker.fancyAuctionHouse.newBid": "新出价：",
     "skyblocker.fancyAuctionHouse.buy": "买！",
