+++ src/main/resources/assets/skyblocker/lang/zh_tw.json
@@ -35,10 +35,10 @@
     "skyblocker.option.locations.barn.solveHungryHiker": "飢餓的旅行者所需物品提示",
     "skyblocker.option.messages.hideMoltenWave": "隱藏Molten Wave技能的提示訊息",
     "skyblocker.option.messages.hideImplosion": "隱藏Implosion技能的提示訊息",
-    "skyblocker.option.richPresence.info.PURSE": "錢包",
-    "skyblocker.option.richPresence.info.LOCATION": "位置",
-    "skyblocker.option.richPresence.info.@Tooltip": "如果您正在循環模式，這個值將不會生效",
-    "skyblocker.option.richPresence.customMessage": "自訂訊息",
+    "skyblocker.option.misc.richPresence.info.PURSE": "錢包",
+    "skyblocker.option.misc.richPresence.info.LOCATION": "位置",
+    "skyblocker.option.misc.richPresence.info.@Tooltip": "如果您正在循環模式，這個值將不會生效",
+    "skyblocker.option.misc.richPresence.customMessage": "自訂訊息",
     "skyblocker.option.locations.dwarvenMines.solveFetchur": "解開Fetchur的謎題",
     "skyblocker.option.messages.chatFilterResult.ACTION_BAR": "移至動作欄",
     "skyblocker.option.messages.chatFilterResult.PASS": "關閉",
@@ -158,7 +158,7 @@
     "skyblocker.option.locations.dungeons.lividColor.enableLividColorText": "啟用真Livid的顏色提示文本",
     "skyblocker.option.locations.dungeons.lividColor.enableLividColorGlow": "啟用Livid高亮",
     "skyblocker.option.general.teleportOverlay.enableSinrecallTransmission": "啟用Sinrecall Transmission技能目標位置顯示",
-    "skyblocker.category.richPresence": "Discord活動狀態",
+    "skyblocker.option.misc.richPresence": "Discord活動狀態",
     "skyblocker.option.locations.dungeons.dungeonChestProfit.enableProfitCalculator.@Tooltip": "在寶箱畫面的標題中顯示地牢寶箱的利潤。\n如果獲利則顯示綠色。\n如果虧損則為紅色。\n灰色表示沒有獲利也沒有虧損。\n如果計算是基於不完整的數據，則顯示為藍色。",
     "skyblocker.option.locations.dungeons.dungeonChestProfit.includeKismet": "包含Kismet價格",
     "skyblocker.option.slayer.endermanSlayer": "終界使者杀手",
@@ -343,7 +343,7 @@
     "skyblocker.option.locations.dwarvenMines.dwarvenHud.screen": "矮人礦井HUD配置…",
     "skyblocker.tips.enabled": "§a啟用提示。",
     "skyblocker.option.quickNav.button": "按鈕 %d",
-    "skyblocker.option.general.betterPartyFinder": "更好的組隊查找",
+    "skyblocker.option.general.fancyPartyFinder": "更好的組隊查找",
     "skyblocker.option.locations.dungeons.secretWaypoints.enableDefaultWaypoints": "啟用預設路徑點",
     "skyblocker.fairySouls.markAllFound": "§r將目前島嶼上的全部仙女之魂標記為已發現",
     "skyblocker.option.locations.dungeons.dungeonScore.enableScoreHUD": "啟用分數狀態欄",
@@ -354,7 +354,7 @@
     "skyblocker.option.locations.dungeons.mimicMessage.sendMimicMessage": "啟用Mimic訊息",
     "skyblocker.option.locations.dungeons.mimicMessage.sendMimicMessage.@Tooltip": "殺死Mimic後在聊天中發送一則訊息，以供其他玩家的分數計算模組使用。",
     "skyblocker.option.locations.dungeons.mimicMessage.mimicMessage": "Mimic訊息",
-    "skyblocker.option.locations.dungeons.lividColor": "提示真Livid的顏色",
+    "skyblocker.option.locations.dungeons.livid": "提示真Livid的顏色",
     "skyblocker.option.locations.dungeons.mimicMessage.mimicMessage.@Tooltip": "殺死Mimic後將在聊天中發送的訊息。 建議保留預設值。",
     "skyblocker.option.locations.dungeons.fireFreezeStaffTimer.@Tooltip": "在F3/M3的Boss戰中使用受到開火凍結影響的武器時顯示計時器。",
     "skyblocker.option.locations.dwarvenMines": "矮人礦井",
@@ -372,7 +372,7 @@
     "skyblocker.option.locations.dwarvenMines.crystalsHud": "水晶洞窟地圖",
     "skyblocker.option.general.enableTips": "啟用貼士",
     "skyblocker.option.general.teleportOverlay.enableTeleportOverlays": "啟用傳送類型技能目標位置顯示",
-    "skyblocker.option.richPresence.enableRichPresence": "已啟用",
+    "skyblocker.option.misc.richPresence.enableRichPresence": "已啟用",
     "skyblocker.option.locations.dwarvenMines.crystalsWaypoints.enabled.@Tooltip": "為叢林遺跡和仙女石窟等水晶洞窟中的重要區域添加路徑點（在一般設定/路徑點中所選擇的路徑點）。 ",
     "skyblocker.option.general.searchOverlay.enableBazaar.@Tooltip": "在Bazaar中搜尋時顯示自訂搜尋介面。",
     "skyblocker.option.general.searchOverlay": "搜尋介面",
@@ -387,7 +387,7 @@
     "skyblocker.option.locations.dungeons.solveBoulder.@Tooltip": "繪製路線並高亮按鈕",
     "skyblocker.option.locations.dungeons.lividColor.enableLividColorTitle": "啟用Livid顏色標題",
     "skyblocker.option.locations.dwarvenMines.crystalsWaypoints.findInChat": "在聊天欄中搜尋路徑點",
-    "skyblocker.option.general.dungeonQuality": "地牢物品品質",
+    "skyblocker.option.general.itemTooltip.dungeonQuality": "地牢物品品質",
     "skyblocker.option.general.searchOverlay.historyLength.@Tooltip": "介面顯示的搜尋歷史的最大數量。",
     "skyblocker.option.general.searchOverlay.enableCommands": "從指令打開",
     "skyblocker.tips.issues": "在 https://github.com/SkyblockerMod/Skyblocker 提交錯誤報告和功能請求。",
@@ -405,8 +405,8 @@
     "skyblocker.option.general.itemTooltip.enableMotesPrice.@Tooltip": "顯示Rift中的物品價格，以Mote為計量單位。",
     "skyblocker.option.general.tabHud.tabHudEnabled": "啟用更好的Tab HUD",
     "skyblocker.option.general.teleportOverlay": "傳送類型技能目標位置顯示",
-    "skyblocker.option.richPresence.info": "Skyblock訊息",
-    "skyblocker.option.richPresence.cycleMode": "循環Skyblock訊息",
+    "skyblocker.option.misc.richPresence.info": "Skyblock訊息",
+    "skyblocker.option.misc.richPresence.cycleMode": "循環Skyblock訊息",
     "skyblocker.updateRepository.error": "§c更新本機資料儲存庫失敗，請手動刪除檔案並重新啟動遊戲。",
     "skyblocker.dungeons.secrets.markSecretFoundUnable": "§c無法將秘密 #%d 標記為已找到。",
     "skyblocker.dungeons.secrets.markSecretMissingUnable": "§c無法將秘密 #%d 標記為已忽略。",
@@ -472,7 +472,7 @@
     "skyblocker.partyFinder.deList": "點擊取消列出",
     "skyblocker.option.general.tabHud.enableHudBackground": "啟用HUD背景",
     "skyblocker.option.general.tabHud.enableHudBackground.@Tooltip": "為非TAB的HUD啟用背景。",
-    "skyblocker.option.richPresence.info.BITS": "點數",
+    "skyblocker.option.misc.richPresence.info.BITS": "點數",
     "skyblocker.option.locations.crimsonIsle.kuudra.safeSpotWaypoints": "安全點路徑",
     "skyblocker.option.locations.crimsonIsle.kuudra.noArrowPoisonWarning": "箭毒耗盡警告",
     "skyblocker.option.locations.garden.farmingHud.enableHud": "啟用耕作HUD",
@@ -530,9 +530,9 @@
     "skyblocker.fancyAuctionHouse.alreadyTopBid": "已經是最高出價！",
     "skyblocker.fancyAuctionHouse.yourAuction": "這是你的拍賣！",
     "text.skyblocker.confirm": "確認",
-    "skyblocker.option.general.betterAuctionHouse.enabled": "啟用",
-    "skyblocker.option.general.betterAuctionHouse": "更好的拍賣行介面",
-    "skyblocker.option.general.betterAuctionHouse.highlightUnderAvgPrice": "高亮顯示低於平均價格的BIN物品",
+    "skyblocker.option.general.fancyAuctionHouse.enabled": "啟用",
+    "skyblocker.option.general.fancyAuctionHouse": "更好的拍賣行介面",
+    "skyblocker.option.general.fancyAuctionHouse.highlightUnderAvgPrice": "高亮顯示低於平均價格的BIN物品",
     "skyblocker.fancyAuctionHouse.price": "價格：",
     "skyblocker.fancyAuctionHouse.newBid": "新出價：",
     "skyblocker.fancyAuctionHouse.buy": "買！",
